converting to monero v7
divide nonce

revise to config file

Updated readme

Updated contacts

bitbucket links -> github

delete to manual ping

skein runtime memory corruption checking

support to force cl_khr_int64_base_atomics

change to stratum protocol (difficulty)

add to config file in cmake

source refactoring (prehash, diff zero count)

change to parse nonce logic

change to difficulty length

delete unnecessary parsing

apply modified target

apply modified format, delete legacy code

set variant as 1

apply job id for each block

refactoring protocol define

chage to icon & remove donate

change to ico

Update README.md

+++ src/amd/GpuContext.h
@@ -6,6 +6,7 @@
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
  * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
  * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright 2018      Team-Hycon  <https://github.com/Team-Hycon>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -97,7 +98,7 @@ struct GpuContext
     int computeUnits;
     std::string name;
 
-    uint32_t Nonce;
+    uint64_t Nonce;
 };
