+++ src/common/net/SubmitResult.h
@@ -6,6 +6,8 @@
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
  * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
  * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright 2018      Team-Hycon  <https://github.com/Team-Hycon>
+ *
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -38,7 +40,7 @@ public:
 
     int64_t reqId;
     int64_t seq;
-    uint32_t diff;
+    uint64_t diff;
     uint64_t actualDiff;
     uint64_t elapsed;
